pyproject.toml
[tool.poetry]
name = "blasttools"
version = "0.1.12"
description = "Commands for turning blast queries into pandas dataframes."
authors = ["Ian Castleden <ian.castleden@uwa.edu.au>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.1"
click = "^8.1.7"
biopython = "^1.81"

[tool.poetry.scripts]
blasttools = "blasttools.__main__:blast"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.0"
pre-commit = "^3.5.0"
jupyter = "^1.0.0"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
